feat: initial release
This commit encapsulates all the work in building and writing the first
version of the site as one big-bang style commit.

This introduces all the technology and dependencies, and ports content
over from https://after.coffee as well as new content.

There is likely to be bugs and things to fix, but this should at least
be a working and mostly correct website for https://gregdaynes.com